requirements.txt
torch==2.0.0
torchvision==0.15.0
scikit-learn==1.3.0
pandas==2.0.0
matplotlib==3.7.0
tqdm==4.66.1
Pillow==10.0.0